+++ src/test/testutil.h
@@ -8,4 +8,8 @@
 #ifndef BITCOIN_TEST_TESTUTIL_H
 #define BITCOIN_TEST_TESTUTIL_H
 
+#include <boost/filesystem/path.hpp>
+
+boost::filesystem::path GetTempPath();
+
 #endif // BITCOIN_TEST_TESTUTIL_H
